create ddf from redshift
