+++ core/src/main/java/io/ddf/DDFManager.java
@@ -28,8 +28,6 @@ import io.ddf.datasource.DataFormat;
 import io.ddf.datasource.DataSourceDescriptor;
 import io.ddf.datasource.DataSourceManager;
 import io.ddf.datasource.SQLDataSourceDescriptor;
-import io.ddf.ds.DataSourceCredential;
-import io.ddf.ds.User;
 import io.ddf.etl.IHandleSqlLike;
 import io.ddf.exception.DDFException;
 import io.ddf.misc.ALoggable;
@@ -654,5 +652,5 @@ public abstract class DDFManager extends ALoggable implements IDDFManager, IHand
     return (new DataSourceManager()).load(ds, this);
   }
 
-  public abstract DDF createDDF(User user, Map<Object, Object> options) throws DDFException;
+  public abstract DDF createDDF(Map<Object, Object> options) throws DDFException;
 }
