+++ core/src/main/java/io/ddf/ds/UserRegistry.java
@@ -15,8 +15,6 @@ public class UserRegistry {
 
   private static final UserRegistry instance = new UserRegistry();
 
-  private static final ThreadLocal<User> currentUser = new ThreadLocal<>();
-
   private final LoadingCache<String, User> nameMap = CacheBuilder.newBuilder().build(
       new CacheLoader<String, User>() {
         @Override
@@ -35,14 +33,6 @@ public class UserRegistry {
 
   public static UserRegistry getInstance() {
     return instance;
-  }
-
-  public static User getCurrentUser() {
-    return currentUser.get();
-  }
-
-  public static void setCurrentUser(User user) {
-    currentUser.set(user);
   }
 
   public synchronized void add(User user) {
