+++ spark/src/main/java/io/ddf/spark/DelegatingDDFManager.java
@@ -2,9 +2,6 @@ package io.ddf.spark;
 
 import io.ddf.DDF;
 import io.ddf.DDFManager;
-import io.ddf.ds.DataSourceCredential;
-import io.ddf.ds.User;
-import io.ddf.ds.UsernamePasswordCredential;
 import io.ddf.exception.DDFException;
 
 import java.util.Map;
@@ -51,13 +48,9 @@ public class DelegatingDDFManager extends DDFManager {
   }
 
   @Override
-  public DDF createDDF(User user, Map<Object, Object> options) throws DDFException {
+  public DDF createDDF(Map<Object, Object> options) throws DDFException {
     options.put("sourceUri", uri);
-    DDF ddf = manager.createDDF(user, options);
-    if (!uri.startsWith("sfdc:")) {
-      ddf.getRepresentationHandler().cache(false);
-    }
-    return ddf;
+    return manager.createDDF(options);
   }
 
   @Override
